linux: xenbus param wait-nonessentials
Add a module parameter to skip waiting up to 30s for non-essential
devices to attach to a backend (vfb, vkbd).

Guests may include the support for different graphic/input frontends and
it might be desirable to conditionally provide and/or attach a backend
to the frontend without waiting for the timeout.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>