openxtfb: blacklist xen-fbfront
openxtfb depends on:
- FB_SYS_FILLRECT
- FB_SYS_COPYAREA
- FB_SYS_IMAGEBLIT
- FB_SYS_FOPS
- FB_DEFERRED_IO
- FB_MODE_HELPERS

In order to select the above configuration, UIVM is built with
XEN_FBDEV_FRONTEND as it shares and selects most of these. Both drivers
will therefor load creating /dev/fb0 and /dev/fb1 by default. This is
undesirable as it makes the XOrg configuration more complicated, so
avoid loading xen-fbfront.

This could be handled differently, e.g, in the xf86 driver.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>