kmeta: split exports from video quirks
openxt-video-quirks applies the patches required for xenfb2 and enables
it. Xenfb2 exports userland headers from the module recipe (which could
be changed to a xenfb2-module-headers like argo-module-headers) which
surfman depends upon in dom0, yet xenfb2 is not supposed to be enabled
in dom0.

Split the exports of internal kernel functions (a hack in itself) and
the actual patch/enabling of the xenfb2 driver to accommodate this. This
way xenfb2 can be built as an out-of-tree module for dom0, export
headers and still not be enabled in the dom0 kernel configuration.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>