installer/answer: package custom answerfiles
package the custom answerfiles provided by the layer.
When using externalsrc and rmwork, the tmpdir where the answer files are
placed by do_unpacked, used in do_deploy, will be wiped. This can lead
to issues when the recipe is used targeting an image that needs
do_deploy to re-run under externalsrc since do_unpack is not always
re-run. Placing the files in the package tracked by the sstate seems to
resolve this issue.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>