Add image-qa-module-sigs.bbclass
image-qa-module-sigs.bbclass checks the signatures of modules
installed in an image to ensure they are valid for the associated
kernel.  Usually this is /boot/bzImage, but it will check in deploy dir
for stubdomain's external kernel.

This adds ~7 seconds to dom0 image creation, but it's nicer to check during
the build than have an install that fails to boot properly.

Tie it in to openxt-image so the check is always performed.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>