layer: amend module-signing dependencies
do_sign_modules will use the STAGING_KERNEL_BUILDDIR, hence add the
depedency to the kernel's shared workdir.

module-signing.bbclass and kernel-module-signing.bbclass let a
pre-generated certificate be given via configuration file:
Add KERNEL_MODULE_SIG_CERT to the vardeps to re-sign modules if that
variable happens to change. Should KERNEL_MODULE_SIG_CERT be set, also
add the targeted file to file-checksum so modules are re-signed if the
certificate is re-generated.

This most likely makes the sstate irrelevant for kernels and out-of-tree
modules. The private key is (and should) not be tracked by the sstate or
packaged, so linux-openxt:do_configure will have to run in derivative
trees, and so its dependent steps.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>