txt_info: add as a patch config fragment
Hosting txt_info.c as an out-of-tree module with its sources in the
layer is not very elegant. The changes are small and configurable, and
useful only to dom0.

Make the out-of-tree module a patch against the kernel and create the
associated configuration fragment so dom0's kernel can select it.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>